+++ tests/unit/request-batcher.js
@@ -654,9 +654,10 @@ describe(`RequestBatcher`, function() {
 
       clock.tick(80000);
       expect(batcher.sendRequest).to.have.been.calledOnce;
-      expect(batcher.sendRequest.args[0][0]).to.deep.equal([
-        {'event': `orphaned event 1`, 'properties': {'foo': 'bar'}},
-      ]);
+      const payload = batcher.sendRequest.args[0][0];
+      expect(payload).to.have.lengthOf(1);
+      expect(payload[0]).to.have.property(`event`, `orphaned event 1`);
+      expect(payload[0]).to.have.nested.include({'properties.foo': `bar`});
 
       expect(getLocalStorageItems()).to.have.lengthOf(2);
       sendResponse(200);
@@ -729,23 +730,25 @@ describe(`RequestBatcher`, function() {
     it(`drops malformed individual items in the localStorage queue`, function() {
       localStorage.setItem(LOCALSTORAGE_KEY, JSON.stringify([
         {id: `fakeID1`, flushAfter: Date.now() - 10000, payload: {
-            'event': `orphaned event 1`, 'properties': {'foo': 'bar'},
+          'event': `orphaned event 1`, 'properties': {'foo': 'bar'},
         }},
 
         `HOW DID THIS RANDOM STRING GET IN MY QUEUE??`,
 
         {id: `fakeID2`, flushAfter: Date.now() - 10000, payload: {
-            'event': `orphaned event 2`,
+          'event': `orphaned event 2`,
         }}
       ]));
       expect(JSON.parse(localStorage.getItem(LOCALSTORAGE_KEY))).to.have.lengthOf(3);
 
       batcher.start();
       expect(batcher.sendRequest).to.have.been.calledOnce;
-      expect(batcher.sendRequest.args[0][0]).to.deep.equal([
-        {'event': `orphaned event 1`, 'properties': {'foo': 'bar'}},
-        {'event': `orphaned event 2`},
-      ]);
+      const payload = batcher.sendRequest.args[0][0];
+      expect(payload).to.have.lengthOf(2);
+      expect(payload[0]).to.have.property(`event`, `orphaned event 1`);
+      expect(payload[0]).to.have.nested.include({'properties.foo': `bar`});
+      expect(payload[1]).to.have.property(`event`, `orphaned event 2`);
+      expect(payload[1]).not.to.have.property(`properties`);
 
       sendResponse(200);
       expect(getLocalStorageItems()).to.be.empty; // invalid item got cleared